pyproject.toml
[tool.poetry]
name = "site-to-lead"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
load-dotenv = "^0.1.0"
loguru = "^0.7.2"
fast-bitrix24 = "^1.7.3"
fastapi = "^0.111.0"
pydantic = "^2.7.3"
requests = "^2.32.3"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
